feat: useStyleState hook
