+++ src/layout/Header/Header.tsx
@@ -1,18 +1,20 @@
 import React from 'react'
-import HeaderState from './HeaderState'
 import MobileHeader from '@/layout/Header/MobileHeader'
 import MobileMenu from '@/layout/Header/MobileMenu'
+import MobileDetailMenu from '@/layout/Header/MobileDetailMenu'
+import StyleStates from '@/features/styleStates'
 
 /** @name {Header} */
 function Header() {
-  const whiteHeaderStyles = 'peer-[.is-white]/header-color:text-base-100'
-  const blackHeaderStyles = 'peer-[.is-black]/header-color:text-base-600'
+  const whiteHeaderStyles = 'peer-[.is-white]/style-state:text-base-100'
+  const blackHeaderStyles = 'peer-[.is-black]/style-state:text-base-600'
 
   return (
     <>
-      <HeaderState />
+      <StyleStates />
       <MobileHeader className={`${whiteHeaderStyles} ${blackHeaderStyles}`} />
       <MobileMenu />
+      <MobileDetailMenu />
     </>
   )
 }
